refactor: substituição do nome antigo e adição do header/footer
